plugins: seperate `Spam Shield` from `Spam Prediction`
